+++ x5crop/detection/outer/proposal/__init__.py
@@ -1,0 +1,3 @@
+"""Outer proposal generators."""
+
+__all__: tuple[str, ...] = ()